requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
backoff==1.11.1
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.10
cryptography==35.0.0
distlib==0.3.4
filelock==3.4.2
frozenlist==1.2.0
future==0.18.2
gcloud-aio-auth==3.7.0
gcloud-aio-pubsub==4.5.0
identify==2.4.4
idna==3.3
multidict==5.2.0
nodeenv==1.6.0
platformdirs==2.4.1
pre-commit==2.16.0
pycparser==2.21
PyJWT==2.3.0
PyYAML==6.0
six==1.16.0
toml==0.10.2
virtualenv==20.13.0
yarl==1.7.2
